create a home page, for a better organization
